Add new functionality to new algorithm
This is means that if we introduce any problems into new
LoadNexusProcessed, it will be easy to move back to version 1.
